requirements.txt
requests
PyYAML
Pillow
pygame
